Propose license update
Hey @mridehalgh! We / k6 recommend changing the license for xk6 from AGPL-3 to Apache 2.0 because we want to assure users that they are free to execute extensions in all modes: locally, using k6-operator on their own infrastructure and in the k6-cloud. Apache 2.0 license makes this more explicit. We know we previously suggested AGPL, sorry for adding work!
There's a bit in the license where you should put in your name or nick (up to you), I marked it as `TODO`.

Thank you!